requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.24.89
botocore==1.27.96
crepe==0.0.16
dlib==19.24.1
fer==22.5.1
groq==0.5.0
imutils==0.5.4
librosa==0.10.2.post1
moviepy==1.0.3
numpy==1.25.2
opencv-python==4.8.0.76
pydub==0.25.1
python-dotenv==1.0.1
scipy==1.11.2
